+++ amethyst_network/Cargo.toml
@@ -31,4 +31,6 @@ bytes = "0.5"
 laminar = "0.3"
 log = "0.4"
 thread_profiler = { version = "0.3" , optional = true }
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
 tungstenite = { version = "0.10.1", optional = true }
